Rename fxa-client crate errors.rs to error.rs
