+++ components/fxa-client/src/config.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::errors::*;
+use crate::error::*;
 use serde_derive::*;
 use std::{cell::RefCell, sync::Arc};
 use url::Url;
