+++ components/fxa-client/src/error.rs
