+++ components/fxa-client/src/http_client.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::{config::Config, errors::*};
+use crate::{config::Config, error::*};
 use serde_derive::*;
 use serde_json::json;
 use std::collections::HashMap;
