+++ components/fxa-client/src/http_client/browser_id.rs
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use crate::{
-    errors::*,
+    error::*,
     http_client::{self, OAuthTokenResponse},
     util::Xorable,
     Config,
