+++ components/fxa-client/src/http_client/browser_id/jwt_utils.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::{errors::*, http_client::browser_id::BrowserIDKeyPair};
+use crate::{error::*, http_client::browser_id::BrowserIDKeyPair};
 use serde_json::{self, json};
 use std::time::{SystemTime, UNIX_EPOCH};
