+++ components/fxa-client/src/scoped_keys.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::{errors::*, FirefoxAccount};
+use crate::{error::*, FirefoxAccount};
 use byteorder::{BigEndian, ByteOrder};
 use ring::{aead, agreement, agreement::EphemeralPrivateKey, digest, rand::SecureRandom};
 use serde_derive::*;
