+++ components/fxa-client/src/util.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::errors::*;
+use crate::error::*;
 use ring::rand::SecureRandom;
 use std::time::{SystemTime, UNIX_EPOCH};
